Adding Stats to UserProfile to start styling
