+++ backend/src/main/java/com/gw/backend/SecurityConfig.java
@@ -1,30 +1,30 @@
-package com.gw.backend;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-import static org.springframework.security.config.Customizer.withDefaults;
-
-@Configuration
-public class SecurityConfig {
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .cors(withDefaults())
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/api/**").authenticated()
-                        .anyRequest().permitAll()
-                )
-                .formLogin(Customizer.withDefaults())
-                .logout(logout -> logout
-                        .logoutUrl("/logout")
-                        .logoutSuccessUrl("/login?logout")
-                        .invalidateHttpSession(true)
-                        .clearAuthentication(true)
-                        .deleteCookies("JSESSIONID")
-                );
-        return http.build();
-    }
-
-}
+//package com.gw.backend;
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.security.config.Customizer;
+//import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+//import org.springframework.security.web.SecurityFilterChain;
+//import static org.springframework.security.config.Customizer.withDefaults;
+//
+//@Configuration
+//public class SecurityConfig {
+//    @Bean
+//    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+//        http
+//                .cors(withDefaults())
+//                .authorizeHttpRequests(auth -> auth
+//                        .requestMatchers("/api/**").authenticated()
+//                        .anyRequest().permitAll()
+//                )
+//                .formLogin(Customizer.withDefaults())
+//                .logout(logout -> logout
+//                        .logoutUrl("/logout")
+//                        .logoutSuccessUrl("/login?logout")
+//                        .invalidateHttpSession(true)
+//                        .clearAuthentication(true)
+//                        .deleteCookies("JSESSIONID")
+//                );
+////        return http.build();
+//    }
+//
+//}
