+++ frontend/src/components/pages/Login.jsx
@@ -1,37 +1,38 @@
 import React from "react";
-import Button from 'react-bootstrap/Button';
-import Form from 'react-bootstrap/Form';
+import Button from "react-bootstrap/Button";
+import Form from "react-bootstrap/Form";
 import { useAuth } from "../auth/AuthContext";
 export default function Login(params) {
-  const {login} = useAuth();
+    const { login } = useAuth();
 
-  const handleLogin = async (e) => {
-    e.preventDefault();
-    login();
-    
-  }
+    const handleLogin = async (e) => {
+        e.preventDefault();
+        login();
+    };
 
-  return (
-    <>
-      <h1>Login</h1>
-      <Form onSubmit={handleLogin}>
-        <Form.Group className="mb-3" controlId="formBasicEmail">
-          <Form.Label>Email address</Form.Label>
-          <Form.Control type="email" placeholder="Enter email" />
-        </Form.Group>
+    return (
+        <>
+            <h1>Login</h1>
+            <Form onSubmit={handleLogin}>
+                <Form.Group className="mb-3" controlId="formBasicEmail">
+                    <Form.Label>Email address</Form.Label>
+                    <Form.Control type="email" placeholder="Enter email" />
+                </Form.Group>
 
-        <Form.Group className="mb-3" controlId="formBasicPassword">
-          <Form.Label>Password</Form.Label>
-          <Form.Control type="password" placeholder="Password" />
-        </Form.Group>
-        <Form.Group className="mb-3" controlId="formBasicCheckbox">
-          <Form.Check type="checkbox" label="Check me out" />
-        </Form.Group>
-        <Button variant="primary" type="submit">
-          Submit
-        </Button>
-        <a href="/sign-up" style={{display:"block"}}>No account? Create one</a>
-      </Form>
-    </>
-  );
+                <Form.Group className="mb-3" controlId="formBasicPassword">
+                    <Form.Label>Password</Form.Label>
+                    <Form.Control type="password" placeholder="Password" />
+                </Form.Group>
+                <Form.Group className="mb-3" controlId="formBasicCheckbox">
+                    <Form.Check type="checkbox" label="Check me out" />
+                </Form.Group>
+                <Button variant="primary" type="submit">
+                    Submit
+                </Button>
+                <a href="/sign-up" style={{ display: "block" }}>
+                    No account? Create one
+                </a>
+            </Form>
+        </>
+    );
 }
